strspn_alt implemented

+++ shell.h
@@ -40,7 +40,7 @@ int strcmp_alt(char *s1, char *s2);
 char *strchr_alt(char *s, char c);
 char *strdup_func(const char *s);
 char *strncat_alt(char *dest, char *src, int n);
-int strspn_alt(char *s, char *accept);
+size_t strspn_alt(char *s, char *accept);
 int strncmp_alt(const char *s1, const char *s2, int n);
 char *strcpy_alt(char *destination, char *source);
 char *strncpy_alt(char *destination, char *source, int n);
